fix win errors move to libusb

+++ picross/src/pic_thread_win32.cpp
@@ -185,7 +185,7 @@ void *pic_thread_lck_malloc(unsigned size)
 
 	// Prepare for locking. - NEED TO SORT QUOTA
 	DWORD pid = GetCurrentProcessId();
-	DWORD sz_min_curr, sz_max_curr;
+	SIZE_T sz_min_curr, sz_max_curr;
 	unsigned long min,max;
 	unsigned long inc = 1024 * 1024 * 20;
